Split TestRunner off of TestScheduler
